chor: refactor auth functions to their own file
